Factor out requests and invites
