+++ shared/teams/team/container.js
@@ -138,8 +138,8 @@ const mergeProps = (stateProps, dispatchProps, ownProps) => {
     ...ownProps,
     customComponent,
     headerStyle: {borderBottomWidth: 0},
-    invites: stateProps._invites.toJS(),
-    requests: stateProps._requests.toJS(),
+    numInvites: stateProps._invites.size,
+    numRequests: stateProps._requests.size,
     newTeamRequests: stateProps._newTeamRequests.toArray(),
     onAddPeople,
     onAddSelf,
